Create contact.php
The contacts page was made -
the contact information shows  a phone number, email address, and address.
Users can provide their name, phone number, email address, nationality, and a message using the contact form. The form has mandatory fields and validation.
Facebook ,Instagram, Twitter, and LinkedIn links are included, along with indivual icon placeholders a rating slider has been added with a slider range of 1 to 10 that allows consumers to rate the service and a button which Contains options to Reset and Submit user interactions okay.