Home Page
I have created the html code for the home page, in this home page what I have completed is the navigation bar, the main page content, which includes 4 categories for now, which are coats, trainers, shirts and joggers. I have also then completed the code for the footer. Completely on the whole, the home page html is finished. Further enhanced implementation of the home page will be coded in.