basket page
i have completed the basket page ,All of the product names, prices, quantities, and subtotals are displayed in a table on the page containing the items in the basket. i have included a button to continue to the checkout i have included an area for applying coupons, and a display of the basket totals are all there. Completely stylized in black, white, and dark grey ,my layout is resposive meaning it adjusts to various screen sizes and offerseasy access on desktop and mobile devices.