Products List Page CSS and Filter Function
I have made the CSS cleaner, so that each row of products presents 4 products each time, I have made it so that when you zoom out, flex wrap does not make the products go all over the place.  I have also routed each item that is displayed, to direct the user to the product details page, which accurately displays each items. I have also fixed up the navigation bar placements so that when you zoom out the logo does not move, and fixed the footer icon and text placements.