Header and Search Bar Complete Change
I have completely Changed the layout of the header, this makes it look very clean, with most of the buttons for account and basket being changed to icons now. The search bar is improved, moved to the right made nicer. The logo has been cleaned up with an additional writing added on it